budget-django/pyproject.toml
[project]
name = "budget"
version = "0.1.0"
description = "Budget app (v4) that handles Rabobank transactions without modifying too much"
authors = [
    {name = "Timo Hermans",email = "tmhermans@gmail.com"}
]
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=6.0.4,<7.0.0)",
    "mozilla-django-oidc (>=5.0.2,<6.0.0)",
    "django-environ (>=0.13.0,<0.14.0)",
    "mypy (>=1.20.1,<2.0.0)",
    "ruff (>=0.15.11,<0.16.0)",
    "django-stubs[compatible-mypy] (>=6.0.2,<7.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "gunicorn (>=25.3.0,<26.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"